third_party/WebKit/LayoutTests/fast/preloader/is-preloaded-after-load-expected.txt

CONSOLE MESSAGE: line 11: PASS: image1.png
CONSOLE MESSAGE: line 12: PASS: non-existant.js
CONSOLE MESSAGE: line 11: PASS: image1.png
CONSOLE MESSAGE: line 12: PASS: non-existant.js
internals.isPreloaded() should remain the same even after ResourceFetcher::clearPreloads() is called (around document's load event). crbug.com/643621

<img src=resources/image1.png>
